refactor: 폴더 구조를 컨벤션에 맞게 정리
